Converted README to rST.
The initial conversion was done using `pandoc`, with manual reformatting done afterwards.

Since both github and PyPI can parse .rst's, it makes sense to keep the documentation in this format rather than markdown.

I also realized that the README included in the source distribution can be a symlink.  So now `README.rst` will be the unique documentation file, and README will be a symlink to it.

Changes were made to `setup.py` that should cause the README to be displayed on the PyPI page.